+++ Assets/Scripts/Networking/Packages/Voyager/MissingFramesRequestPacket.cs
@@ -1,0 +1,7 @@
+﻿namespace VoyagerApp.Networking.Packages.Voyager
+{
+    public class MissingFramesRequestPacket : Packet
+    {
+        public MissingFramesRequestPacket() : base(OpCode.MissingFramesRequest) { }
+    }
+}